app/requirements.txt
fastapi
uvicorn
openai>=1.0.0
pyodbc
pandas
python-dotenv
azure-identity
azure-keyvault-secrets